Add delete purge artifact empty artifact

+++ server/tasks/purgeArtifact.ts
@@ -1,4 +1,4 @@
-import { lt } from "drizzle-orm";
+import { isNull, lt, ne } from "drizzle-orm";
 import db from "../db/db";
 import { S3Fetch } from "../services/s3fetch";
 
@@ -15,11 +15,22 @@ export default defineTask({
             ...env,
             enableLogging: false,
         })
-        const purgeAppArtifact = await drizzle.select().from(tables.purgeAppArtifact)
-            .orderBy(desc(tables.purgeAppArtifact.createdAt))
-            .limit(1)
-            .then(singleOrDefault)
-        if (!purgeAppArtifact) {
+        const [purgeAppArtifact, purgeAppNonArtifact] = await Promise.all([
+            drizzle.select().from(tables.purgeAppArtifact)
+                .orderBy(desc(tables.purgeAppArtifact.createdAt))
+                .where(eq(tables.purgeAppArtifact.hasArtifact, 1))
+                .limit(1)
+                .then(singleOrDefault),
+            drizzle.select().from(tables.purgeAppArtifact)
+                .orderBy(asc(tables.purgeAppArtifact.hasArtifact))
+                .where(or(
+                    ne(tables.purgeAppArtifact.hasArtifact, 1),
+                    isNull(tables.purgeAppArtifact.hasArtifact),
+                ))
+                .limit(1)
+                .then(singleOrDefault),
+        ])
+        if (!purgeAppArtifact && !purgeAppNonArtifact) {
             return {
                 result: {
                     status: 'nothing to purge',
@@ -27,18 +38,36 @@ export default defineTask({
             }
         }
 
-        await drizzle.transaction(async tx => {
-            const hasArtifact = purgeAppArtifact.hasArtifact === 1 ? true : false
-            if (hasArtifact) {
-                console.log('Purge has artifact, will check storage', purgeAppArtifact.appId)
-                throw 'purging storage not implemented'
-            }
+        async function purgeArtifactInternal() {
+            if (purgeAppArtifact) {
+                await drizzle.transaction(async tx => {
+                    const hasArtifact = purgeAppArtifact.hasArtifact === 1 ? true : false
+                    if (hasArtifact) {
+                        console.log('Purge has artifact, will check storage', purgeAppArtifact.appId)
+                        const storageOrgPrefix = getStorageKeysOrg(purgeAppArtifact.orgId!, purgeAppArtifact.appId!)
+                        throw 'purging storage not implemented, keys ' + storageOrgPrefix
+                    }
 
-            await tx.delete(tables.purgeAppArtifact)
-                .where(and(
-                    eq(tables.purgeAppArtifact.orgId, purgeAppArtifact.orgId!),
-                    eq(tables.purgeAppArtifact.appId, purgeAppArtifact.appId!),
-                ))
+                    await tx.delete(tables.purgeAppArtifact)
+                        .where(and(
+                            eq(tables.purgeAppArtifact.orgId, purgeAppArtifact.orgId!),
+                            eq(tables.purgeAppArtifact.appId, purgeAppArtifact.appId!),
+                        ))
+                })
+            }
+        }
+        async function purgeNonArtifactInternal() {
+            if (purgeAppNonArtifact) {
+                await drizzle.delete(tables.purgeAppArtifact)
+                    .where(and(
+                        eq(tables.purgeAppArtifact.orgId, purgeAppNonArtifact.orgId!),
+                        eq(tables.purgeAppArtifact.appId, purgeAppNonArtifact.appId!),
+                    ))
+            }
+        }
+        const results = await Promise.allSettled([purgeArtifactInternal(), purgeNonArtifactInternal()])
+        results.forEach(e => {
+            console.log('Result purge artifact', e)
         })
 
         console.log('Finished purge artifact')
